+++ demo/src/app/components/alert/demos/selfclosing/alert-selfclosing.ts
@@ -1,27 +1,28 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, OnInit, ViewChild} from '@angular/core';
 import {Subject} from 'rxjs';
 import {debounceTime} from 'rxjs/operators';
+import {NgbAlert} from '@ng-bootstrap/ng-bootstrap';
 
-@Component({
-  selector: 'ngbd-alert-selfclosing',
-  templateUrl: './alert-selfclosing.html'
-})
+@Component({selector: 'ngbd-alert-selfclosing', templateUrl: './alert-selfclosing.html'})
 export class NgbdAlertSelfclosing implements OnInit {
   private _success = new Subject<string>();
 
   staticAlertClosed = false;
   successMessage = '';
 
+  @ViewChild('staticAlert', {static: false}) staticAlert: NgbAlert;
+  @ViewChild('selfClosingAlert', {static: false}) selfClosingAlert: NgbAlert;
+
   ngOnInit(): void {
-    setTimeout(() => this.staticAlertClosed = true, 20000);
+    setTimeout(() => this.staticAlert.close(), 20000);
 
     this._success.subscribe(message => this.successMessage = message);
-    this._success.pipe(
-      debounceTime(5000)
-    ).subscribe(() => this.successMessage = '');
+    this._success.pipe(debounceTime(5000)).subscribe(() => {
+      if (this.selfClosingAlert) {
+        this.selfClosingAlert.close();
+      }
+    });
   }
 
-  public changeSuccessMessage() {
-    this._success.next(`${new Date()} - Message successfully changed.`);
-  }
+  public changeSuccessMessage() { this._success.next(`${new Date()} - Message successfully changed.`); }
 }