+++ demo/src/app/components/progressbar/demos/height/progressbar-height.ts
@@ -2,12 +2,7 @@ import {Component} from '@angular/core';
 
 @Component({
   selector: 'ngbd-progressbar-height',
-  templateUrl: './progressbar-height.html',
-  styles: [`
-    ngb-progressbar {
-      margin-top: 5rem;
-    }
-  `]
+  templateUrl: './progressbar-height.html'
 })
 export class NgbdProgressbarHeight {
   height = '20px';