+++ demo/src/app/components/progressbar/demos/labels/progressbar-labels.ts
@@ -2,12 +2,7 @@ import {Component} from '@angular/core';
 
 @Component({
   selector: 'ngbd-progressbar-labels',
-  templateUrl: './progressbar-labels.html',
-  styles: [`
-    ngb-progressbar {
-      margin-top: 5rem;
-    }
-  `]
+  templateUrl: './progressbar-labels.html'
 })
 export class NgbdProgressbarLabels {
 }