+++ demo/src/app/components/tabset/tabset.module.ts
@@ -26,43 +26,50 @@ const DEMOS = {
     title: 'Tabset',
     type: NgbdTabsetBasic,
     code: require('!!raw-loader!./demos/basic/tabset-basic').default,
-    markup: require('!!raw-loader!./demos/basic/tabset-basic.html').default
+    markup: require('!!raw-loader!./demos/basic/tabset-basic.html').default,
+    showStackblitz: false
   },
   pills: {
     title: 'Pills',
     type: NgbdTabsetPills,
     code: require('!!raw-loader!./demos/pills/tabset-pills').default,
-    markup: require('!!raw-loader!./demos/pills/tabset-pills.html').default
+    markup: require('!!raw-loader!./demos/pills/tabset-pills.html').default,
+    showStackblitz: false
   },
   selectbyid: {
     title: 'Select an active tab by id',
     type: NgbdTabsetSelectbyid,
     code: require('!!raw-loader!./demos/selectbyid/tabset-selectbyid').default,
-    markup: require('!!raw-loader!./demos/selectbyid/tabset-selectbyid.html').default
+    markup: require('!!raw-loader!./demos/selectbyid/tabset-selectbyid.html').default,
+    showStackblitz: false
   },
   preventchange: {
     title: 'Prevent tab change',
     type: NgbdTabsetPreventchange,
     code: require('!!raw-loader!./demos/preventchange/tabset-preventchange').default,
-    markup: require('!!raw-loader!./demos/preventchange/tabset-preventchange.html').default
+    markup: require('!!raw-loader!./demos/preventchange/tabset-preventchange.html').default,
+    showStackblitz: false
   },
   justify: {
     title: 'Nav justification',
     type: NgbdTabsetJustify,
     code: require('!!raw-loader!./demos/justify/tabset-justify').default,
-    markup: require('!!raw-loader!./demos/justify/tabset-justify.html').default
+    markup: require('!!raw-loader!./demos/justify/tabset-justify.html').default,
+    showStackblitz: false
   },
   orientation: {
     title: 'Nav orientation',
     type: NgbdTabsetOrientation,
     code: require('!!raw-loader!./demos/orientation/tabset-orientation').default,
-    markup: require('!!raw-loader!./demos/orientation/tabset-orientation.html').default
+    markup: require('!!raw-loader!./demos/orientation/tabset-orientation.html').default,
+    showStackblitz: false
   },
   config: {
     title: 'Global configuration of tabs',
     type: NgbdTabsetConfig,
     code: require('!!raw-loader!./demos/config/tabset-config').default,
-    markup: require('!!raw-loader!./demos/config/tabset-config.html').default
+    markup: require('!!raw-loader!./demos/config/tabset-config.html').default,
+    showStackblitz: false
   }
 };