+++ demo/src/app/components/toast/demos/howto-global/toasts-container.component.ts
@@ -11,7 +11,7 @@ import {ToastService} from './toast-service';
       [class]="toast.classname"
       [autohide]="true"
       [delay]="toast.delay || 5000"
-      (hide)="toastService.remove(toast)"
+      (hidden)="toastService.remove(toast)"
     >
       <ng-template [ngIf]="isTemplate(toast)" [ngIfElse]="text">
         <ng-template [ngTemplateOutlet]="toast.textOrTpl"></ng-template>