+++ src/carousel/carousel-config.ts
@@ -1,4 +1,5 @@
 import {Injectable} from '@angular/core';
+import {NgbConfig} from '../ngb-config';
 
 /**
  * A configuration service for the [NgbCarousel](#/components/carousel/api#NgbCarousel) component.
@@ -15,4 +16,11 @@ export class NgbCarouselConfig {
   pauseOnFocus = true;
   showNavigationArrows = true;
   showNavigationIndicators = true;
+
+  private _animation: boolean;
+
+  constructor(private _ngbConfig: NgbConfig) {}
+
+  get animation(): boolean { return (this._animation === undefined) ? this._ngbConfig.animation : this._animation; }
+  set animation(animation: boolean) { this._animation = animation; }
 }