+++ src/test/ngb-config-animation.ts
@@ -1,0 +1,7 @@
+import {Injectable} from '@angular/core';
+import {NgbConfig} from '../ngb-config';
+
+@Injectable()
+export class NgbConfigAnimation extends NgbConfig {
+  animation = true;
+}